backend/requirements.txt
asgiref==3.8.1
Django==3.2.12
django-cors-headers==3.8.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
mssql-django==1.4.2
PyJWT==2.8.0
pyodbc==5.1.0
pytz==2024.1
sqlparse==0.5.0
typing-extensions==4.11.0